requirements.txt
echo "streamlit>=1.40.0
python-dotenv>=0.19.0
pillow>=9.0.0
twelvelabs>=1.0.0
requests>=2.28.0" > requirements.txt